Main2

+++ app/src/main/java/fiap/com/br/fiapapp/model/Veiculo.kt
@@ -9,7 +9,7 @@ data class Veiculo (
     var cod_marca: Int? = -1,
     var cod_cor: Int? = -1,
     var cnpj: Long? = -1,
-    var km: Long? = -1L,
+    var km: Int? = -1,
     var valor: Double? = -1.00,
     var ano: Int? = -1,
     var filial: Int? = -1,
